added new favicon. corrected a type error in BitsharesAsset. added WalletBurn to the bitshares rpc library. corrected another type error in delegate_participation. added cors to the http repsonses so we can support localhost requests, or requests from embedded widgets on different hosts. added an emergency suspend to service whenever delegate participation drops below 90%. we can now issue UIAs as well as transfer assets, to support crowd funding.

+++ BitsharesRpc/BitsharesTypes.cs
@@ -47,7 +47,7 @@ namespace BitsharesRpc
 		public string name;
 		public string description;
 		public string public_data;
-		public ulong issuer_account_id;
+		public uint issuer_account_id;
 		public ulong precision;
 		public DateTime registration_date;
 		public DateTime last_update;
